test: 객체별로 테스트 분리
